+++ addon/components/flexberry-checkbox.js
@@ -23,7 +23,7 @@ export default FlexberryBaseComponent.extend({
     @property value
     @type Boolean
     @default false
-   */
+  */
   value: false,
 
   /**
@@ -60,6 +60,15 @@ export default FlexberryBaseComponent.extend({
   tagName: '',
 
   /**
+    DOM-element representing semantic ui-checkbox component.
+
+    @property checkboxDomElement
+    @type Object
+    @default null
+  */
+  checkboxDomElement: null,
+
+  /**
     Path to component's settings in application configuration (JSON from ./config/environment.js).
 
     @property appConfigSettingsPath
@@ -76,4 +85,32 @@ export default FlexberryBaseComponent.extend({
       checked: self.get('value'),
     });
   }),
+
+  /**
+    Called when the element of the view has been inserted into the DOM or after the view was re-rendered.
+    For more information see [didInsertElement](http://emberjs.com/api/classes/Ember.Component.html#event_didInsertElement) event of [Ember.Component](http://emberjs.com/api/classes/Ember.Component.html).
+  */
+  didInsertElement() {
+    this._super(...arguments);
+
+    // We need to select and remember DOM-element representing ui-checkbox component,
+    // but we can't just call to this.$('.flexberry-checkbox'), because ember will throw an error
+    // "You cannot access this.$() on a component with tagName: ''.".
+    // So we have to search our element in 'childViews' collection.
+    let checkboxView = this.get('childViews').find(view => {
+      let tagName = view.get('tagName');
+      if (Ember.typeOf(tagName) === 'string') {
+        tagName = tagName.trim();
+      }
+
+      return tagName !== '' && view.$().hasClass('flexberry-checkbox');
+    });
+
+    let checkboxDomElement = !Ember.isNone(checkboxView) ? checkboxView.$() : null;
+    this.set('checkboxDomElement', checkboxDomElement);
+
+    checkboxDomElement.checkbox({
+      uncheckable: true
+    });
+  },
 });