+++ CHANGELOG.md
@@ -13,7 +13,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
     * Added support of HTML attributes.
 
 ### Fixed
-
+* `flexberry-checkbox` component:
+    * Now if `flexberry-checkbox` was unchecked it doest'n has checked class.
 
 ## [0.4.0] - 2016-08-15
 ### Changed
@@ -37,7 +38,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.3.0] - 2016-07-29
 ### Added
-* Custom internationalized captions for boolean type in `object-list-view-cell` commponent.
+* Custom internationalized captions for boolean type in `object-list-view-cell` component.
 * Blueprints:
     * Add regeneration for models and serializers. Now models and serizlizers generates into separate regeneratable mixins, model and serializer classes will not be replaced during regeneration.
     * Add new `flexberry-group` blueprint for generation of group of entities by one blueprint.